requirements.txt
flask
gunicorn
request
bs4
re